Moving the article's creator FAB
