+++ app/src/back-office/style/home-layout.scss
@@ -59,3 +59,11 @@
   display: flex;
   justify-content: center;
 }
+
+.add-button {
+    bottom: 0;
+    right: 0;
+    margin: 25px;
+    position: fixed;
+    padding-left: 0;
+  }
