+++ app/src/back-office/views/layouts/home-layout.jsx
@@ -1,16 +1,26 @@
 import React, {Component, PropTypes} from 'react';
 import {connect} from 'react-redux';
 import i18n from 'i18next';
+import {saveArticle} from '../../../common/actions/article-detail';
 import {PasswordComponent} from '../../../common/components/password';
 import {showSnackBar} from '../../../common/actions/snack-bar';
-import {Snackbar, Paper, FlatButton} from 'material-ui';
+import {Snackbar, Paper, FlatButton, FloatingActionButton, Dialog, TextField} from 'material-ui';
 import {green500, red500} from 'material-ui/styles/colors';
 import {withRouter} from 'react-router';
 import {appUrl} from '../../../common/server/config';
 
 /** Home-Layout component. */
 @withRouter
-@connect(state => ({snackBar: state.snackBar}))
+@connect(
+    state => ({
+        connected: state.login.isConnected
+    }),
+    dispatch => ({
+        saveArticle: (title, description) => dispatch(saveArticle({title, description})),
+        showSnackBar: data => dispatch(showSnackBar(data))
+    })
+)
+@connect(state => ({snackBar: state.snackBar, connected: state.login.isConnected}))
 export default class HomeLayout extends Component {
 
     static propTypes = {
@@ -20,6 +30,36 @@ export default class HomeLayout extends Component {
     state = {open: false};
 
     togglePopover = (e) => this.setState({open: !this.state.open, anchorEl: e.currentTarget});
+
+    async saveArticle() {
+        const title = this.refs.title.getValue();
+        const description = this.refs.description.getValue();
+        if (title && description) {
+            try {
+                const id = await this.props.saveArticle(title, description);
+                this.props.showSnackBar({
+                    message: 'edit-cartridge.content.snackBar.saveSuccessMessage',
+                    isError: false
+                });
+                this.toggleModal();
+                setTimeout(() => this.props.router.push(`edit-article/${id}`), 500);
+            } catch (e) {
+                this.props.showSnackBar({
+                    message: 'edit-cartridge.content.snackBar.saveFailedMessage',
+                    isError: true
+                });
+            }
+        } else {
+            this.props.showSnackBar({
+                message: 'edit-cartridge.content.snackBar.saveFailedContentMessage',
+                isError: true
+            });
+        }
+    }
+
+    toggleModal() {
+        this.setState({open: !this.state.open});
+    }
 
     render() {
         const {children, Content, snackBar: {show, message, timeout, actionText, actionHandler, isError}} = this.props;
@@ -59,6 +99,24 @@ export default class HomeLayout extends Component {
                     onRequestClose={() => this.props.dispatch(showSnackBar()) }
                     bodyStyle={{backgroundColor: isError ? red500 : green500}}
                     />
+                    {this.props.connected ?
+                <FloatingActionButton onClick={() => this.toggleModal()} className='add-button'>
+                    <i className='material-icons'>add</i>
+                </FloatingActionButton>
+            : null}
+
+                <Dialog
+                        title={i18n.t('article.create.dialog')}
+                        actions={[<FlatButton label={i18n.t('article.create.confirm')} primary={true} onClick={() => this.saveArticle()} />]}
+                        open={this.state.open}
+                        onRequestClose={() => this.toggleModal()}
+                    >
+                    <div>
+                        <TextField ref='title' hintText={i18n.t('article.title')} />
+                        <br />
+                        <TextField ref='description' hintText={i18n.t('article.description')} />
+                    </div>
+                </Dialog>
             </div>
         );
     }
