+++ app/src/common/components/article-list/index.jsx
@@ -2,11 +2,8 @@ import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
 
 import {loadArticleList} from '../../actions/article-list';
-import {saveArticle} from '../../actions/article-detail';
 import {showSnackBar} from '../../actions/snack-bar';
 import {ArticleList as List} from './list';
-import {Dialog, FlatButton, TextField} from 'material-ui';
-import i18n from 'i18next';
 import {withRouter} from 'react-router';
 
 /** Component that displays the list of all articles, connected to the store. */
@@ -18,7 +15,6 @@ import {withRouter} from 'react-router';
     }),
     dispatch => ({
         loadArticleList: () => dispatch(loadArticleList()),
-        saveArticle: (title, description) => dispatch(saveArticle({title, description})),
         showSnackBar: data => dispatch(showSnackBar(data))
     })
 )
@@ -41,53 +37,11 @@ export class ArticleList extends React.Component {
         }
     }
 
-    async saveArticle() {
-        const title = this.refs.title.getValue();
-        const description = this.refs.description.getValue();
-        if (title && description) {
-            try {
-                const id = await this.props.saveArticle(title, description);
-                this.props.showSnackBar({
-                    message: 'edit-cartridge.content.snackBar.saveSuccessMessage',
-                    isError: false
-                });
-                this.toggleModal();
-                setTimeout(() => this.props.router.push(`edit-article/${id}`), 500);
-            } catch (e) {
-                this.props.showSnackBar({
-                    message: 'edit-cartridge.content.snackBar.saveFailedMessage',
-                    isError: true
-                });
-            }
-        } else {
-            this.props.showSnackBar({
-                message: 'edit-cartridge.content.snackBar.saveFailedContentMessage',
-                isError: true
-            });
-        }
-    }
-
-    toggleModal() {
-        this.setState({open: !this.state.open});
-    }
-
     render() {
         const {connected, articleList} = this.props;
         return (
             <div style={{flex: 1}}>
-                <List articleList={articleList} connected={connected} openCreate={() => this.toggleModal()} />
-                <Dialog
-                    title={i18n.t('article.create.dialog')}
-                    actions={[<FlatButton label={i18n.t('article.create.confirm')} primary={true} onClick={() => this.saveArticle()} />]}
-                    open={this.state.open}
-                    onRequestClose={() => this.toggleModal()}
-                >
-                    <div>
-                        <TextField ref='title' hintText={i18n.t('article.title')} />
-                        <br />
-                        <TextField ref='description' hintText={i18n.t('article.description')} />
-                    </div>
-                </Dialog>
+                <List articleList={articleList} connected={connected} />
             </div>
         );
     }
